Rework testcase modules for main entrypoints
